use div instead of svg for toast desktop/tablet background
