+++ app/website-projects/article-preview-card/_components/share-toast-triangle.tsx
@@ -1,0 +1,19 @@
+function ShareToastTriangle() {
+  return (
+    <svg
+      width="24"
+      height="12"
+      viewBox="0 0 24 12"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className="text-grey-900"
+    >
+      <path
+        d="M0 0 L24 0 L12 12 Z"
+        fill="currentColor" // Change fill color as needed
+      />
+    </svg>
+  );
+}
+
+export default ShareToastTriangle;
